+++ proto/job.proto
@@ -29,7 +29,7 @@ message Uuid {
   uint64 id2 = 2;
 }
 
-message Job {
+message JobData {
   Uuid id = 1;
   optional uint64 last_updated = 2;
   optional uint64 last_tick = 3;
