+++ src/job_data.rs
@@ -1,52 +1,52 @@
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct CronJob {
-    #[prost(string, tag = "1")]
+    #[prost(string, tag="1")]
     pub schedule: ::prost::alloc::string::String,
 }
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct NonCronJob {
-    #[prost(bool, tag = "1")]
+    #[prost(bool, tag="1")]
     pub repeating: bool,
-    #[prost(uint64, tag = "2")]
+    #[prost(uint64, tag="2")]
     pub repeated_every: u64,
 }
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct Uuid {
-    #[prost(uint64, tag = "1")]
+    #[prost(uint64, tag="1")]
     pub id1: u64,
-    #[prost(uint64, tag = "2")]
+    #[prost(uint64, tag="2")]
     pub id2: u64,
 }
 #[derive(Clone, PartialEq, ::prost::Message)]
-pub struct Job {
-    #[prost(message, optional, tag = "1")]
+pub struct JobData {
+    #[prost(message, optional, tag="1")]
     pub id: ::core::option::Option<Uuid>,
-    #[prost(uint64, optional, tag = "2")]
+    #[prost(uint64, optional, tag="2")]
     pub last_updated: ::core::option::Option<u64>,
-    #[prost(uint64, optional, tag = "3")]
+    #[prost(uint64, optional, tag="3")]
     pub last_tick: ::core::option::Option<u64>,
-    #[prost(uint64, tag = "4")]
+    #[prost(uint64, tag="4")]
     pub next_tick: u64,
-    #[prost(enumeration = "JobType", tag = "5")]
+    #[prost(enumeration="JobType", tag="5")]
     pub job_type: i32,
-    #[prost(uint32, tag = "8")]
+    #[prost(uint32, tag="8")]
     pub count: u32,
-    #[prost(message, repeated, tag = "9")]
+    #[prost(message, repeated, tag="9")]
     pub on_start: ::prost::alloc::vec::Vec<Uuid>,
-    #[prost(message, repeated, tag = "10")]
+    #[prost(message, repeated, tag="10")]
     pub on_stop: ::prost::alloc::vec::Vec<Uuid>,
-    #[prost(message, repeated, tag = "11")]
+    #[prost(message, repeated, tag="11")]
     pub on_remove: ::prost::alloc::vec::Vec<Uuid>,
-    #[prost(oneof = "job::Job", tags = "6, 7")]
-    pub job: ::core::option::Option<job::Job>,
+    #[prost(oneof="job_data::Job", tags="6, 7")]
+    pub job: ::core::option::Option<job_data::Job>,
 }
-/// Nested message and enum types in `Job`.
-pub mod job {
+/// Nested message and enum types in `JobData`.
+pub mod job_data {
     #[derive(Clone, PartialEq, ::prost::Oneof)]
     pub enum Job {
-        #[prost(message, tag = "6")]
+        #[prost(message, tag="6")]
         CronJob(super::CronJob),
-        #[prost(message, tag = "7")]
+        #[prost(message, tag="7")]
         NonCronJob(super::NonCronJob),
     }
 }
